feat: add collection to SolFrenSDK
